extract ThreeColumnLayout
