+++ src/routes/docs/getting-started/introduction.mdx
@@ -1,4 +1,4 @@
-import { Lead } from "../../docs"
+import { Lead } from "../../../components/ui"
 import Snippet from "../../../components/snippet"
 import { Tabs, Tab } from "../../../components/tabs"
